remove detector types from ast
