+++ source-code/core/src/ast/schema.ts
@@ -2,16 +2,6 @@
 
 export type Language = string
 
-export type Detector = () => Language[] | Language | undefined
-
-export type InitDetector<Parameters extends Array<unknown>> = (
-	...parameters: Parameters
-) => Detector
-
-export type DetectorTemplate<Parameters extends Array<unknown>> = (
-	...parameters: Parameters
-) => ReturnType<Detector>
-
 export type NodeName =
 	| "Identifier"
 	| "Resource"
