+++ source-code/sdk-js/src/detectors/rootSlugDetector.ts
@@ -1,4 +1,5 @@
-import type { Language, InitDetector, DetectorTemplate } from "@inlang/core/ast"
+import type { Language } from "@inlang/core/ast"
+import type { DetectorTemplate, InitDetector } from "./types.js"
 
 type DetectorParameters = [
 	{
